backend-flask/requirements.txt
# Flask and CORS
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

pyjwt
pymongo

# LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langchain-text-splitters>=0.0.1

# Vector database
faiss-cpu>=1.7.4

# PDF processing
pypdf>=3.17.0

# Token counting
tiktoken>=0.5.0

# Word document processing
python-docx>=0.8.11

# SQL database support
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.9

